ajout du delete & update controller pour l'org création du schema, module, controller et service Pld

+++ apps/backend/src/app/organization/organization.service.ts
@@ -1,8 +1,13 @@
 import { Injectable } from '@nestjs/common';
 import {InjectModel} from "@nestjs/mongoose";
 import {Organization, OrganizationDocument} from "./organization.schema";
-import {Model} from "mongoose";
+import {Model, ObjectId} from "mongoose";
 import {CreateOrganizationBody} from "../../../../../libs/data-access/organization/CreateOrganizationBody";
+import {UserDocument} from "../user/user.schema";
+import {UpdateOrganizationBody} from "../../../../../libs/data-access/organization/UpdateOrganizationBody";
+import {DeleteOrganizationBody} from "../../../../../libs/data-access/organization/DeleteOrganizationBody";
+
+export type MemberUpdateObjects = UserDocument[] | string[] | ObjectId[];
 
 @Injectable()
 export class OrganizationService {
@@ -30,13 +35,26 @@ export class OrganizationService {
     return this.organizationModel.create(org);
   }
 
-  public delete(orgObjectId: string): Promise<OrganizationDocument | null> {
-    return this.organizationModel.findOneAndDelete({_id: orgObjectId})
+  public async delete(ownerId: string, orgObjectId: string): Promise<OrganizationDocument | null> {
+    return this.organizationModel.findOneAndDelete({_id: orgObjectId, owner: ownerId})
       .exec();
   }
 
-  public update(orgObjectId: string, org: Organization): Promise<OrganizationDocument | null> {
-    return this.organizationModel.findOneAndUpdate({_id: orgObjectId}, org, {new: true})
+  public async deleteByBody(owner: string | UserDocument, body: DeleteOrganizationBody): Promise<OrganizationDocument | null> {
+    return this.organizationModel.findOneAndDelete({_id: body.orgId, owner}).exec();
+  }
+
+  public update(orgObjectId: string, ownerId: string, org: Organization): Promise<OrganizationDocument | null> {
+    return this.organizationModel.findOneAndUpdate({_id: orgObjectId, owner: ownerId}, org, {new: true})
+      .populate('owner')
+      .populate('members')
+      .exec();
+  }
+
+  public updateByBody(ownerId: string, body: UpdateOrganizationBody) {
+    const orgId = body.orgId;
+    delete body.orgId;
+    return this.organizationModel.findOneAndUpdate({_id: orgId, owner: ownerId}, {...body, updated_date: new Date()}, {new: true})
       .populate('owner')
       .populate('members')
       .exec();
@@ -56,13 +74,13 @@ export class OrganizationService {
       .exec();
   }
 
-  public async addMember(orgId: string, userId: string): Promise<OrganizationDocument | null> {
-    return this.organizationModel.findOneAndUpdate({_id: orgId}, {$addToSet: {members: userId}}, {new: true, populate: ['members', 'owner']})
+  public async addMembers(orgId: string, ownerId: string, userId: MemberUpdateObjects): Promise<OrganizationDocument | null> {
+    return this.organizationModel.findOneAndUpdate({_id: orgId, owner: ownerId}, {$addToSet: {members: userId}}, {new: true, populate: ['members', 'owner']})
       .exec();
   }
 
-  public async removeMember(orgId: string, userId: string): Promise<OrganizationDocument | null> {
-    return this.organizationModel.findOneAndUpdate({_id: orgId}, {$pull: {members: userId}}, {new: true, populate: ['members', 'owner']})
+  public async removeMembers(orgId: string, ownerId:string, userId: MemberUpdateObjects): Promise<OrganizationDocument | null> {
+    return this.organizationModel.findOneAndUpdate({_id: orgId, owner: ownerId}, {$pull: {members: {$in: userId}}}, {new: true, populate: ['members', 'owner']})
       .exec();
   }
